feat : use map method to show all screen items
